Overhaul several types
